feat(table): update table with classmodifier
