+++ src/shared/components/LiveCode/LiveCode.tsx
@@ -2,7 +2,7 @@ import vsDark from 'prism-react-renderer/themes/vsDark';
 import React from 'react';
 import { LiveProviderProps, LiveProvider, LiveError, LivePreview } from 'react-live';
 import * as reactTookitAll from 'shared/toolkit';
-import withClassNameModifier from 'shared/hoc/WithClassNameModifier';
+import withClassNameModifier, { TwithClassNameModifier } from 'shared/hoc/WithClassNameModifier';
 import TabsLiveCode from './TabsLiveCode';
 import Accessibility from './Accessibility';
 import Code from './Code';
@@ -21,7 +21,7 @@ type TLiveCode = {
   hideCode?: boolean;
   hideAccessibility?: boolean;
   hideReadme?: boolean;
-};
+} & TwithClassNameModifier;
 
 const styleLivePreviewDefault = { background: 'white', padding: '2rem', width: '100%' } as const;
 const ariaLabel = 'af-accessibility' as const;
@@ -50,7 +50,7 @@ const LiveCode = withClassNameModifier(
       </TabsLiveCode>
     </article>
   ),
-  'af-livecode',
+  { defaultClassName: 'af-livecode' },
 );
 
 export default LiveCode;
