+++ src/shared/components/Table/Body/Line/Td/__tests__/Td.test.tsx
@@ -4,22 +4,23 @@ import Td from '../Td';
 
 const defaultProps = {
   children: null,
+  label: 'label',
+  hover: null,
 };
 const container = document.createElement('tr');
 
 describe('Td', () => {
-  it('Render <Td/>', () => {
-    const { baseElement } = renderWithContainer(<Td {...defaultProps}>child td</Td>, container);
-    expect(baseElement).toMatchSnapshot();
+  it('Render <Td/> without hover', () => {
+    const { asFragment } = renderWithContainer(<Td {...defaultProps}>child th</Td>, container);
+    expect(asFragment()).toMatchSnapshot();
   });
-
-  it('Render <Td/> with modifier', () => {
-    const { baseElement } = renderWithContainer(
-      <Td {...defaultProps} classModifier="custom">
+  it('Render <Td/> with hover', () => {
+    const { asFragment } = renderWithContainer(
+      <Td {...defaultProps} classModifier="custom" hover={<p>Hover content</p>}>
         child td
       </Td>,
       container,
     );
-    expect(baseElement).toMatchSnapshot();
+    expect(asFragment()).toMatchSnapshot();
   });
 });
