+++ src/shared/components/Table/Header/Th/__tests__/Th.test.tsx
@@ -1,17 +1,49 @@
-import { describe, it, expect } from 'vitest';
-import { renderWithContainer } from 'shared/testsUtils';
-import Th from '../Th';
+import { describe, it, expect, vi } from 'vitest';
+import { render } from '@testing-library/react';
+import { ASCENDING } from 'shared/components/Table/constants';
+import ThContainer from '../Th';
 
-const container = document.createElement('tr');
+const onSort = vi.fn();
+const ThSortableCmpt = vi.fn();
+const ThCmpt = vi.fn();
+const defaultProps = {
+  sorting: {
+    field: '',
+    order: ASCENDING,
+  },
+  field: '',
+  onSort,
+  ThSortableCmpt,
+  ThCmpt,
+} as const;
 
-describe('Th', () => {
-  it('Render <Th/>', () => {
-    const { baseElement } = renderWithContainer(<Th>child th</Th>, container);
-    expect(baseElement).toMatchSnapshot();
+describe('ThContainer', () => {
+  it('Render <ThContainer/> with field empty (not sortable)', () => {
+    render(<ThContainer {...defaultProps}>child th</ThContainer>);
+    expect(ThCmpt).toHaveBeenCalledWith(
+      {
+        children: 'child th',
+        role: 'columnheader',
+      },
+      {},
+    );
   });
 
-  it('Render <Th/> with modifier', () => {
-    const { baseElement } = renderWithContainer(<Th classModifier="custom">child th</Th>, container);
-    expect(baseElement).toMatchSnapshot();
+  it('Render <ThContainer/> with field not empty (sortable)', () => {
+    const customProps = {
+      ...defaultProps,
+      field: 'myfield',
+      sorting: { field: 'name', order: ASCENDING },
+    } as const;
+    render(<ThContainer {...customProps}>child th</ThContainer>);
+    expect(ThSortableCmpt).toHaveBeenCalledWith(
+      {
+        sorting: { field: 'name', order: ASCENDING },
+        field: 'myfield',
+        children: 'child th',
+        onSort,
+      },
+      {},
+    );
   });
 });
